Change unit test with to have correct includes
This fixes some tests that are fine on my desktop but
the subprocess aborts even when all tests are green in the CI.
The error is this: malloc_consolidate(): invalid chunk size
Was told this was a heap corruption. And changing the
includes fixes the error, not sure why.
Changed also to only use one instance of CCDB api
so we do not call CURL global cleanup out of order
with the BasicCCDBManger singelton. Just in case
that matters when running the test(s).